hotFix:add ArticleCommentComponent
